# radix51: use go1.12 intrinsics for 128-bit multiplications