--- a/internal/radix51/fe_test.go
+++ b/internal/radix51/fe_test.go
@@ -9,50 +9,32 @@ import (
 	"crypto/rand"
 	"io"
 	"testing"
-	"unsafe"
 )
 
 func TestMul64to128(t *testing.T) {
 	a := uint64(5)
 	b := uint64(5)
-	r0, r1 := mul64x64(0, 0, a, b)
+	r0, r1 := madd64(0, 0, a, b)
 	if r0 != 0x19 || r1 != 0 {
 		t.Errorf("lo-range wide mult failed, got %d + %d*(2**64)", r0, r1)
 	}
 
 	a = uint64(18014398509481983) // 2^54 - 1
 	b = uint64(18014398509481983) // 2^54 - 1
-	r0, r1 = mul64x64(0, 0, a, b)
+	r0, r1 = madd64(0, 0, a, b)
 	if r0 != 0xff80000000000001 || r1 != 0xfffffffffff {
 		t.Errorf("hi-range wide mult failed, got %d + %d*(2**64)", r0, r1)
 	}
 
 	a = uint64(1125899906842661)
 	b = uint64(2097155)
-	r0, r1 = mul64x64(0, 0, a, b)
-	r0, r1 = mul64x64(r0, r1, a, b)
-	r0, r1 = mul64x64(r0, r1, a, b)
-	r0, r1 = mul64x64(r0, r1, a, b)
-	r0, r1 = mul64x64(r0, r1, a, b)
+	r0, r1 = madd64(0, 0, a, b)
+	r0, r1 = madd64(r0, r1, a, b)
+	r0, r1 = madd64(r0, r1, a, b)
+	r0, r1 = madd64(r0, r1, a, b)
+	r0, r1 = madd64(r0, r1, a, b)
 	if r0 != 16888498990613035 || r1 != 640 {
 		t.Errorf("wrong answer: %d + %d*(2**64)", r0, r1)
-	}
-}
-
-func BenchmarkWideMultInline(t *testing.B) {
-	var r0, r1, ol, oh uint64
-	a := uint64(18014398509481983) // 2^54 - 1
-	b := uint64(18014398509481983) // 2^54 - 1
-
-	for i := 0; i < t.N; i++ {
-		t1 := (a>>32)*(b&0xFFFFFFFF) + ((a & 0xFFFFFFFF) * (b & 0xFFFFFFFF) >> 32)
-		t2 := (a&0xFFFFFFFF)*(b>>32) + (t1 & 0xFFFFFFFF)
-		ol = (a * b) + r0
-		cmp := ol < r0
-		oh = r1 + (a>>32)*(b>>32) + t1>>32 + t2>>32 + uint64(*(*byte)(unsafe.Pointer(&cmp)))
-
-		r1 = oh
-		r0 = ol
 	}
 }
 
@@ -62,7 +44,7 @@ func BenchmarkWideMultCall(t *testing.B) {
 	b := uint64(18014398509481983)
 
 	for i := 0; i < t.N; i++ {
-		r0, r1 = mul64x64(r0, r1, a, b)
+		r0, r1 = madd64(r0, r1, a, b)
 	}
 }