--- a/internal/radix51/mul_compat.go
+++ b/internal/radix51/mul_compat.go
@@ -2,13 +2,14 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
+// +build !go1.12
+
 package radix51
 
 import "unsafe"
 
-// mul64x64 multiples two 64-bit numbers and adds them to two accumulators.
-// This function is written to ensure it inlines. I am so sorry.
-func mul64x64(lo, hi, a, b uint64) (ol uint64, oh uint64) {
+// madd64 multiples two 64-bit numbers and adds them to a split 128-bit accumulator.
+func madd64(lo, hi, a, b uint64) (ol uint64, oh uint64) {
 	t1 := (a>>32)*(b&0xFFFFFFFF) + ((a & 0xFFFFFFFF) * (b & 0xFFFFFFFF) >> 32)
 	t2 := (a&0xFFFFFFFF)*(b>>32) + (t1 & 0xFFFFFFFF)
 	ol = (a * b) + lo